SceneFunctions van elkaar gescheiden, zodat het makkelijker is om erin te werken.
