+++ HSR01JavaFX/src/Applicatie/Applicatie.java
@@ -1,6 +1,7 @@
 package Applicatie;
 
 import Config.config;
+import Functions.Scenes.LoginScene;
 import javafx.application.Application;
 import javafx.scene.image.Image;
 import javafx.stage.Stage;
@@ -12,7 +13,7 @@ public class Applicatie extends Application {
         stage.setTitle("Pakketmonitor");
         stage.getIcons().add(new Image(config.ICON));
         
-        stage.setScene(SceneFunction.loginScene(stage));
+        stage.setScene(LoginScene.getScene(stage));
         
         stage.setResizable(false);
         stage.show();
