--- HSR01JavaFX/src/Classes/Afzender.java
@@ -1,23 +1,0 @@
-package Classes;
-
-
-
-public class Afzender extends Account {
-    //Attributes
-    
-    
-    //Constructors
-    public Afzender(int i, String vn, String tv, String an, String pc, String hnr, String toev, String em, String ww, String gbd, String tnr) {
-        super(i, vn, tv, an, pc, hnr, toev, em, ww, gbd, tnr);
-    }
-    
-    //Getters
-    
-    
-    //Setters
-    
-    
-    //Methods
-    
-    
-}
