--- HSR01JavaFX/src/Classes/Ontvanger.java
@@ -1,23 +1,0 @@
-package Classes;
-
-
-
-public class Ontvanger extends Persoon {
-    //Attributes
-    
-    
-    //Constructors
-    public Ontvanger(int i, String vn, String tv, String an, String pc, String hnr, String toev, String em) {
-        super(i, vn, tv, an, pc, hnr, toev, em);
-    }
-    
-    //Getters
-    
-    
-    //Setters
-    
-    
-    //Methods
-    
-    
-}
