+++ HSR01JavaFX/src/Verouderd/LoginScene.java
@@ -1,6 +1,6 @@
 package Verouderd;
 
-import Functions.DatabaseArrayListHashMap;
+import Functions.Database.DatabaseArrayListHashMap;
 import Functions.Hasher;
 import java.util.ArrayList;
 import java.util.Iterator;
